moving seeds directory back to test directory
