+++ test/integration/clientRoutes.test.js
@@ -10,7 +10,7 @@ const {
   putClientRoute,
   deleteClientRoute,
 } = require('../helper/clientRouteHelper')
-const { cleanupAll, cleanup } = require('../../seeds/project')
+const { cleanupAll, cleanup } = require('../seeds/project')
 
 describe('Client routes', function () {
   let app
